+++ trunk/ diy-layout-creator --username bancika@gmail.com/Library/src/org/diylc/components/semiconductors/TransistorTO92.java
@@ -98,7 +98,7 @@ public class TransistorTO92 extends AbstractTransparentComponent<String> {
 	}
 
 	private void updateControlPoints() {
-		int pinSpacing = PIN_SPACING.convertToPixels();
+		int pinSpacing = (int) PIN_SPACING.convertToPixels();
 		// Update control points.
 		int x = controlPoints[0].x;
 		int y = controlPoints[0].y;
@@ -128,7 +128,7 @@ public class TransistorTO92 extends AbstractTransparentComponent<String> {
 		if (body == null) {
 			int x = controlPoints[0].x;
 			int y = controlPoints[0].y;
-			int pinSpacing = PIN_SPACING.convertToPixels();
+			int pinSpacing = (int) PIN_SPACING.convertToPixels();
 			int bodyDiameter = getClosestOdd(BODY_DIAMETER.convertToPixels());
 
 			switch (orientation) {
@@ -166,7 +166,7 @@ public class TransistorTO92 extends AbstractTransparentComponent<String> {
 	@Override
 	public void draw(Graphics2D g2d, ComponentState componentState, Project project,
 			IDrawingObserver drawingObserver) {
-		int pinSize = PIN_SIZE.convertToPixels() / 2 * 2;
+		int pinSize = (int) PIN_SIZE.convertToPixels() / 2 * 2;
 		Area mainArea = getBody();
 		Composite oldComposite = g2d.getComposite();
 		if (alpha < MAX_ALPHA) {
